Выполнил задание "Отрисуй меня полностью"
