+++ js/main.js
@@ -1,1 +1,5 @@
 import './data.js';
+import {PHOTOS_COUNT, createImage } from './data.js';
+import {createPictures} from './pictures.js';
+
+createPictures(Array.from( {length: PHOTOS_COUNT}, createImage));
